+++ cdsdashboards/cdsalembic/env.py
@@ -61,6 +61,7 @@ def run_migrations_offline():
     context.configure(
         url=url,
         target_metadata=target_metadata,
+        render_as_batch=True,
         literal_binds=True,
         dialect_opts={"paramstyle": "named"},
         version_table=version_table
@@ -86,7 +87,8 @@ def run_migrations_online():
     with connectable.connect() as connection:
         context.configure(
             connection=connection, target_metadata=target_metadata,
-            version_table=version_table
+            version_table=version_table,
+            render_as_batch=True
         )
 
         with context.begin_transaction():