+++ cdsdashboards/hubextension/base.py
@@ -93,38 +93,11 @@ class SpawnPermissionsController():
 
 class DashboardBaseMixin:
 
-    unsafe_regex = re.compile(r'[^a-zA-Z0-9]+')
-
-    trailingdash_regex = re.compile(r'\-+$')
-
     name_regex = re.compile(r'^[a-zA-Z0-9][a-zA-Z0-9\._\- \/~\:\!\@\$#\[\]\(\)\&\*\+\?\<\>\.\'\" ,;\%\=]{0,99}$')
 
     conda_env_regex = re.compile(r'^[a-zA-Z0-9\._\- \/~\:\!\@\$#\[\]\(\)\&\*\+\?\<\>\.\'\" ,;\%\=]+$')
 
     start_path_regex = re.compile(r'^[A-Za-z0-9\-\._~\:\/\?#\[\]@\!\$\&\(\)\*\+\'\" ,;\%\=]*$')
-
-    def calc_urlname(self, dashboard_name):
-        base_urlname = re.sub(self.unsafe_regex, '-', dashboard_name).lower()[:35]
-
-        base_urlname = re.sub(self.trailingdash_regex, '', base_urlname)
-
-        urlname = base_urlname
-
-        self.log.debug('calc safe name from '+urlname)
-
-        now_unique = False
-        counter = 1
-        while not now_unique:
-            orm_dashboard = Dashboard.find(db=self.db, urlname=urlname)
-            self.log.info("{} - {}".format(urlname, orm_dashboard))
-            if orm_dashboard is None or counter >= 100:
-                now_unique = True
-            else:
-                urlname = "{}-{}".format(base_urlname, counter)
-                counter += 1
-
-        self.log.debug('calc safe name : '+urlname)
-        return urlname
 
     @staticmethod
     async def pipe_spawner_progress(dashboard_user, new_server_name, builder):
@@ -156,9 +129,6 @@ class DashboardBaseMixin:
                         pass
 
                 break 
-
-    def get_source_spawners(self, user):
-        return [spawner_to_dict(spawner) for spawner in user.all_spawners(include_default=True) if not spawner.orm_spawner.dashboard_final_of]
 
     def get_visitor_dashboards(self, user):
         orm_dashboards = set()
@@ -269,26 +239,6 @@ class DashboardBaseMixin:
 
             if existing_build_future:
                 await existing_build_future
-
-            # This section would be better if it was inside dockerbuilder, but we want 
-            # to make use of self.spawn_single_user without repeating the code
-            if dashboard.source_spawner and dashboard.source_spawner.name:
-
-                # Get source spawner
-                source_spawner = dashboard_user.spawners[dashboard.source_spawner.name]
-
-                if source_spawner.ready:
-                    self.log.debug('Source spawner is already ready')
-                elif source_spawner.pending in ['spawn', 'check']:
-                    builder.add_progress_event({'progress': 15, 'message': 'Attaching to pending source server for Dashboard'})
-                    self.log.debug('Source spawner is pending - await')
-                    spawn_future = getattr(final_spawner, '_{}_future'.format(final_spawner.pending), None)
-                    if spawn_future:
-                        await spawn_future
-                else:
-                    builder.add_progress_event({'progress': 10, 'message': 'Starting up source server for Dashboard'})
-                    self.log.debug('Source spawner needs a full start')
-                    await self.spawn_single_user(dashboard_user, dashboard.source_spawner.name)
 
             # Delete existing final spawner if it exists
             await self.maybe_delete_existing_server(dashboard.final_spawner, dashboard_user)