+++ examples/local_process_spawner.py
@@ -121,12 +121,14 @@ c.VariableMixin.extra_presentation_launchers = {
 
 
 c.CDSDashboardsConfig.show_source_git = True
-c.CDSDashboardsConfig.show_source_servers = False
 
 c.CDSDashboardsConfig.conda_envs = ['', 'env1', 'env2', 'cds']
 #c.CDSDashboardsConfig.allow_custom_conda_env = True
 
 #c.CDSDashboardsConfig.default_allow_all = True
+
+#c.CDSDashboardsConfig.spawn_as_viewer = True
+#c.CDSDashboardsConfig.jupyter_startpath_regex = "^shared/"
 
 #c.VariableMixin.voila_template = 'materialstream'